requirements.txt
pyyaml~=5.3.1
Image~=1.5.33
lib_platform~=1.2.7